refactor: Rename Vault Icon
