+++ src/widgets/Menu/icons/Vault.tsx
