+++ src/widgets/Menu/icons/index.ts
@@ -1,4 +1,4 @@
-export { default as BalleVaultIcon } from './BalleVault';
+export { default as VaultIcon } from './Vault';
 export { default as FarmIcon } from './Farm';
 export { default as GroupsIcon } from './Groups';
 export { default as HamburgerIcon } from './Hamburger';
